reorganize to allow all buttons to work as expected
