+++ app/components/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router';
-import _ from 'underscore';
 
 export default class App extends Component {
 
@@ -16,17 +15,17 @@ export default class App extends Component {
     })
   }
 
-  filterFavorites (dbFavs, newFavs) {
-    let formatted = newFavs.map((stuff) => {
-      return stuff.action.newFav
-    })
-    let finalFaves = (dbFavs).concat(formatted) || []
-    let noDuplicates = _.uniq(finalFaves, (movie) => {
-      return movie.title;
-    });
-    this.props.setFinalFavs(noDuplicates);
-    this.props.switchToFavs();
-  }
+  // filterFavorites (dbFavs, newFavs) {
+  //   let formatted = newFavs.map((stuff) => {
+  //     return stuff.action.newFav
+  //   })
+  //   let finalFaves = (dbFavs).concat(formatted) || []
+  //   let noDuplicates = _.uniq(finalFaves, (movie) => {
+  //     return movie.title;
+  //   });
+  //   this.props.setFinalFavs(noDuplicates);
+  //   // this.props.switchToFavs();
+  // }
 
   render() {
     return (
@@ -35,12 +34,6 @@ export default class App extends Component {
           <h1>Not Quite Netflix</h1>
           <Link to='/login'>
             {!this.props.userSignInReducer.user ? <button className='signIn'> Sign In </button> : ''}
-          </Link>
-          <Link to={'/favorites'} >
-            {this.props.userSignInReducer.user && !this.props.movieReducer.toggle ? <button className='favs' onClick={() => this.filterFavorites(this.props.userSignInReducer.fav.data.data, this.props.movieIndexReducer)}> Show Favorites </button> : ''}
-          </Link>
-          <Link to={'/'} >
-            {this.props.movieReducer.toggle ? <button className='back' > Back </button> : ''}
           </Link>
         </div>
         {this.props.children}
