+++ app/components/FavMovies.js
@@ -1,10 +1,17 @@
 import React, {Component} from 'react';
+import {Link} from 'react-router';
 import MovieIndex from '../containers/MovieIndex-container';
 
 export default class FavMovies extends Component{
+
   render(){
     return(
-      <MovieIndex movies={this.props.movieReducer.final.finalFaves}/>
+      <div>
+        <Link to={'/'} >
+          <button className='back'> Back </button>
+        </Link>
+        <MovieIndex movies={this.props.allMoviesReducer.finalFaves}/>
+      </div>
     )
   }
 }
