+++ app/reducers/index.js
@@ -2,12 +2,14 @@ import {combineReducers} from 'redux';
 import movieReducer from './reducer-movie';
 import userSignInReducer from './reducer-userSignIn';
 import movieIndexReducer from './reducer-movieIndex';
+import allMoviesReducer from './reducer-allMovies';
 
 
 const allReducers = combineReducers({
   movieReducer,
   userSignInReducer,
   movieIndexReducer,
+  allMoviesReducer,
 })
 
 export default allReducers;
