+++ app/reducers/reducer-allMovies.js
@@ -1,0 +1,9 @@
+const allMoviesReducer = (state = [], action) => {
+  switch(action.type){
+    case 'FINAL_FAVS':
+      return action;
+  }
+  return state;
+}
+
+export default allMoviesReducer;
