+++ app/reducers/reducer-movie.js
@@ -2,8 +2,6 @@ const movieReducer = (state = [], action) => {
   switch(action.type){
     case 'ADD_MOVIES':
       return Object.assign({}, state, {movies: action.movies})
-    case 'FINAL_FAVS':
-      return Object.assign({}, state, {final: action})
     case 'TOGGLE_FAVS':
       return Object.assign({}, state, {toggle: action})
   }
